first stab at loading message history
Turns out that client-side message management is more complex
than the server side.

This commit has message history kinda work, although it still
leaves undefined a lot of behavior related to client reconnection.

Additionally, I haven't figured out yet what the correct memory
model for messages is, which I think is also impacting my ability
to wire history fetching to UI scrolling, hence the crappy
'fetch history' button :^)

This commit ported over to the new db system a couple of queries
that were left behind and fixes an important bug regarding message
reference counting when more than one client is connected.

Related to the above, this commit introduces '-Dlocal-cache', a
flag for awebo-client that makes it store `.cache/` and `.config/`
in cwd, which is useful when trying to connect from two different
accounts at the same time.

+++ src/awebo/protocol/media.zig
@@ -83,7 +83,7 @@ pub const Header = extern struct {
 /// An open stream request is marked by Packet Sequence Number set to 0.
 /// The rest of the header data will be ignored.
 pub const OpenStream = extern struct {
-    tcp_client: u64,
+    tcp_client: i128,
     nonce: u64,
 
     //  0                   1                   2                   3
